Added usable Nuspec editor and Pack/Push
The Nuspec Editor has been more flushed out and can be used to save,
load, and edit .nuspec files.  It still doesn't support everything, such
as dependencies.

The Nuget Helper was updated to support the Pack and Push commands to
create .nupkg files and export them to the server.

The Nuspec Editor makes use of the Pack and Push commands.