Added a new HTTP-based List method
Added a new method that gets a list of packages on the server.  Instead
of using nuget.exe, this uses the HTTP interface of NuGet.Server and
NuGet Gallery which is less error-prone and has more features.

This required adding the System.ServiceModel.Web.dll in order to get
access to the SyndicationFeed object to easily read the NuGet OData/Atom
feed.