Added preliminary version range checking
It is completely untested.

It follows these rules:
If the package, or a later version, is already installed, use it.
If an earlier version is installed, update it.
If not installed, look on the server for specific version
If specific version not found on server, use the next version up (not
latest)