Fixed bug with cleaning
The Clean operation was incorrectly checking the ENTIRE directory path
instead of just the final directory name for the names of directories to
delete.

This resulted in a bug where folders were deleted if the path contained
any of the names of directories to delete.

It now correctly only checks the final directory name.