Files/Folders starting with . are now packed
The -NoDefaultExcludes option was added to the "nuget pack" command to
allow files and folders that start with a . to be packed into the
package.

This was done because if you want a file/folder in a Unity project, but
you want Unity to ignore it, it must start with a .

This is especially useful for .cs and .js files that you don't want
Unity to compile as game scripts.