tools folder is installed OUTSIDE the Assets folder
If a package contains a tools folder, it is now installed outside of the
Assets folder in the Unity project (alongside Library, Assets, and
ProjectSettings).

This allows tool scripts, and all of their dependent binaries, to be
installed without worrying about trying to load those binaries as
assemblies.  This is due to those binaries possibly not being compatible
with Unity or its version of Mono that it uses.

This folder is also deleted when the package is uninstalled.