send password with HTTP request, enable .NET 4.6
This failed to download packages from VSTS when using .NET 3.5, returning an error concerning the authorization header. I wasn't able to track down the cause, but it seems to be something that is fixed when using the .NET 4.6 framework. This framework will become the default in a forthcoming version of Unity.